feat: move color-scheme into lib
