--- color-scheme/.eslintrc
@@ -1,77 +1,0 @@
-{
-  "root": true,
-  "parser": "@typescript-eslint/parser", // Specifies the ESLint parser
-  "parserOptions": {
-    "ecmaFeatures": {
-      "jsx": true,
-      "impliedStrict": true
-    },
-    "ecmaVersion": 2020, // Allows for the parsing of modern ECMAScript features
-    "sourceType": "module", // Allows for the use of imports
-    "tsconfigRootDir": "./"
-  },
-  "env": {
-    "es6": true,
-    "browser": true,
-    "node": true
-  },
-  "settings": {
-    "react": {
-      "version": "detect"
-    },
-    "import/resolver": {
-      "node": {
-        "extensions": [".js", ".jsx", ".ts", ".tsx"]
-      },
-      "typescript": {
-        "alwaysTryTypes": true
-      }
-    }
-  },
-  "extends": [
-    "eslint:recommended",
-    "plugin:@typescript-eslint/recommended",
-    "plugin:import/errors",
-    "plugin:import/warnings",
-    "plugin:jsx-a11y/strict",
-    "plugin:react/recommended",
-    "prettier/@typescript-eslint",
-    "prettier"
-  ],
-  "plugins": [
-    "simple-import-sort",
-    "react-hooks",
-    "@typescript-eslint",
-    "import"
-  ],
-  "rules": {
-    "no-prototype-builtins": "off",
-    "no-unneeded-ternary": ["warn", { "defaultAssignment": false }],
-    "no-unused-vars": "off",
-    "@typescript-eslint/no-unused-vars": [
-      "warn",
-      { "ignoreRestSiblings": true }
-    ],
-    "no-useless-concat": "warn",
-    "no-useless-escape": "warn",
-    "prefer-const": "error",
-    "prefer-template": "warn",
-    "react-hooks/exhaustive-deps": "warn",
-    "react-hooks/rules-of-hooks": "error",
-    "react/display-name": "off",
-    "react/jsx-boolean-value": "warn",
-    "react/jsx-curly-brace-presence": ["warn", { "props": "never" }],
-    "react/jsx-fragments": "warn",
-    "react/jsx-no-useless-fragment": "error",
-    "react/no-array-index-key": "warn",
-    "react/no-did-update-set-state": "warn",
-    "react/no-direct-mutation-state": "error",
-    "react/no-unescaped-entities": "warn",
-    "react/prefer-stateless-function": "warn",
-    "react/prop-types": "off",
-    "react/self-closing-comp": "warn",
-    "simple-import-sort/imports": "warn",
-    "jsx-a11y/label-has-for": "off", // Turning this off as it is deprecated, using label-has-associated-control instead
-    "jsx-a11y/label-has-associated-control": "warn"
-  }
-}
