--- color-scheme/.lintstagedrc
@@ -1,7 +1,0 @@
-{
-  "src/**/*.{js,ts,tsx}": [
-    "eslint --fix --cache --quiet",
-    "prettier --config .prettierrc --write"
-  ],
-  "*.json": "prettier --config .prettierrc --write"
-}
