--- color-scheme/.prettierignore
@@ -1,3 +1,0 @@
-dist
-node_modules
-coverage
