--- color-scheme/.prettierrc
@@ -1,4 +1,0 @@
-trailingComma: "none"
-tabWidth: 2
-semi: false
-singleQuote: true
