--- color-scheme/package.json
@@ -1,52 +1,0 @@
-{
-  "name": "@atom-learning/color-scheme",
-  "source": "src/index.ts",
-  "main": "dist/index.cjs.js",
-  "typings": "dist/index.d.ts",
-  "module": "dist/index.js",
-  "version": "0.0.0-semantically-released",
-  "scripts": {
-    "dev": "NODE_ENV=development rollup -c -w",
-    "build": "NODE_ENV=production rollup -c",
-    "clean": "del ./dist/",
-    "format": "prettier 'src/**/*.{js,ts,tsx}' --write",
-    "lint": "eslint 'src/**/*.{js,ts,tsx}' --fix"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/Atom-Learning/components.git"
-  },
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "@rollup/plugin-node-resolve": "^11.2.0",
-    "@typescript-eslint/eslint-plugin": "^4.29.1",
-    "@typescript-eslint/parser": "^4.29.1",
-    "esbuild": "^0.14.10",
-    "eslint": "^7.18.0",
-    "eslint-config-prettier": "^7.2.0",
-    "eslint-import-resolver-typescript": "^2.3.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-react": "^7.22.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "eslint-plugin-simple-import-sort": "^7.0.0",
-    "prettier": "^2.2.1",
-    "rollup": "^2.39.0",
-    "rollup-plugin-json": "^4.0.0",
-    "rollup-plugin-esbuild": "^4.8.2",
-    "rollup-plugin-visualizer": "^4.2.2",
-    "typescript": "4.2.2"
-  },
-  "peerDependencies": {
-    "@atom-learning/icons": "^1",
-    "@atom-learning/theme": "^1",
-    "@types/react": "^16 || ^17 || ^18",
-    "react": "^16 || ^17 || ^18",
-    "react-dom": "^16 || ^17 || ^18"
-  },
-  "dependencies": {
-    "@stitches/react": "^1.2.8",
-    "color2k": "^2.0.0"
-  }
-}
