+++ lib/package.json
@@ -126,7 +126,6 @@
     "react-dom": "^16.0.0 || ^17.0.0 || ^18.0.0"
   },
   "dependencies": {
-    "@atom-learning/color-scheme": "0.0.0-semantically-released",
     "@atom-learning/theme": "1.1.1",
     "@dnd-kit/core": "^6.0.5",
     "@dnd-kit/modifiers": "^6.0.0",
