+++ lib/src/components/tabs/TabsTriggerList.tsx
@@ -1,10 +1,10 @@
 import { List } from '@radix-ui/react-tabs'
 import React from 'react'
 import { styled } from '~/stitches'
-import { ColorScheme } from '@atom-learning/color-scheme'
-import { useWindowSize } from '../../utilities/hooks/useWindowSize'
-import { useScrollPosition } from '../../utilities/hooks/useScrollPosition'
-import { useCallbackRefState } from '../../utilities/hooks/useCallbackRef'
+import { ColorScheme } from '~/experiments/color-scheme'
+import { useWindowSize } from '~/utilities/hooks/useWindowSize'
+import { useScrollPosition } from '~/utilities/hooks/useScrollPosition'
+import { useCallbackRefState } from '~/utilities/hooks/useCallbackRef'
 import { ActionIcon } from '../action-icon'
 import { Icon } from '../icon'
 import { ChevronLeft, ChevronRight } from '@atom-learning/icons'
