+++ lib/src/experiments/color-scheme/blue.json
