+++ lib/src/experiments/color-scheme/slate.json
